+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/LogWatchCallback.java
@@ -16,7 +16,6 @@
 package io.fabric8.kubernetes.client.dsl.internal;
 
 import io.fabric8.kubernetes.client.KubernetesClientException;
-import io.fabric8.kubernetes.client.RequestConfigBuilder;
 import io.fabric8.kubernetes.client.dsl.LogWatch;
 import io.fabric8.kubernetes.client.http.AsyncBody;
 import io.fabric8.kubernetes.client.http.HttpClient;
@@ -32,9 +31,7 @@ import java.net.URL;
 import java.nio.ByteBuffer;
 import java.nio.channels.Channels;
 import java.nio.channels.WritableByteChannel;
-import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 public class LogWatchCallback implements LogWatch, AutoCloseable {
@@ -43,7 +40,6 @@ public class LogWatchCallback implements LogWatch, AutoCloseable {
 
   private final OutputStream out;
   private WritableByteChannel outChannel;
-  private final OperationContext context;
   private volatile InputStream output;
 
   private final AtomicBoolean closed = new AtomicBoolean(false);
@@ -55,7 +51,6 @@ public class LogWatchCallback implements LogWatch, AutoCloseable {
     if (out != null) {
       outChannel = Channels.newChannel(out);
     }
-    this.context = context;
     this.serialExecutor = new SerialExecutor(context.getExecutor());
   }
 
@@ -74,14 +69,10 @@ public class LogWatchCallback implements LogWatch, AutoCloseable {
 
   public LogWatchCallback callAndWait(HttpClient client, URL url) {
     HttpRequest request = client.newHttpRequestBuilder().url(url).build();
-    HttpClient clone = client.newBuilder()
-        .tag(Optional.ofNullable(context.getRequestConfig()).map(RequestConfigBuilder::new).orElse(new RequestConfigBuilder())
-            .withRequestTimeout(0).build())
-        .readTimeout(0, TimeUnit.MILLISECONDS).build();
 
     if (out == null) {
       // we can pass the input stream directly to the consumer
-      clone.sendAsync(request, InputStream.class).whenComplete((r, e) -> {
+      client.sendAsync(request, InputStream.class).whenComplete((r, e) -> {
         if (e != null) {
           onFailure(e);
         }
@@ -92,7 +83,7 @@ public class LogWatchCallback implements LogWatch, AutoCloseable {
     } else {
       // we need to write the bytes to the given output
       // we don't know if the write will be blocking, so hand it off to another thread
-      clone.consumeBytes(request, (buffers, a) -> CompletableFuture.runAsync(() -> {
+      client.consumeBytes(request, (buffers, a) -> CompletableFuture.runAsync(() -> {
         for (ByteBuffer byteBuffer : buffers) {
           try {
             outChannel.write(byteBuffer);