+++ kubernetes-client/src/test/java/io/fabric8/kubernetes/client/dsl/internal/PortForwarderWebsocketTest.java
@@ -34,7 +34,7 @@ class PortForwarderWebsocketTest {
 
   @BeforeEach
   void initPortForwarderWebsocket() {
-    this.portForwarderWebsocket = new PortForwarderWebsocket(mockHttpClient, CommonThreadPool.get());
+    this.portForwarderWebsocket = new PortForwarderWebsocket(mockHttpClient, CommonThreadPool.get(), 0);
   }
 
   @Test